requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
langchain-pinecone
langchain-huggingface
pinecone[grpc]
sentence-transformers
python-dotenv
pypdf
-e .